chore: expose maptos state view.
