+++ Cargo.toml
@@ -84,6 +84,7 @@ movement-client = { git = "https://github.com/movementlabsxyz/movement.git", rev
 aptos-executor= { git = "https://github.com/movementlabsxyz/maptos-core.git", rev = "4e1871804307948d43002fc81d8ea723caeb5c31" }
 aptos-vm = { git = "https://github.com/movementlabsxyz/maptos-core.git", rev = "4e1871804307948d43002fc81d8ea723caeb5c31" }
 aptos-rest-client = { git = "https://github.com/movementlabsxyz/maptos-core.git", rev = "4e1871804307948d43002fc81d8ea723caeb5c31" }
+aptos-storage-interface = { git = "https://github.com/movementlabsxyz/maptos-core.git", rev = "4e1871804307948d43002fc81d8ea723caeb5c31" }
 
 # model checking and verification
