+++ tests/executor/citeria/global-storage-equal/Cargo.toml
@@ -10,6 +10,7 @@ rust-version = { workspace = true }
 
 [dependencies]
 migration-executor-test-types = { workspace = true }
+anyhow = { workspace = true }
 
 
 [lints]
