+++ tests/executor/citeria/global-storage-equal/src/lib.rs
@@ -2,9 +2,9 @@ use migration_executor_test_types::criterion::{
 	Criterion, CriterionError, Criterionish, MaptosExecutor, MovementExecutor,
 };
 
-pub struct Empty;
+pub struct GlobalStorageEqual;
 
-impl Empty {
+impl GlobalStorageEqual {
 	pub fn new() -> Self {
 		Self
 	}
@@ -14,12 +14,24 @@ impl Empty {
 	}
 }
 
-impl Criterionish for Empty {
+impl Criterionish for GlobalStorageEqual {
 	fn satisfies(
 		&self,
 		movement_executor: &MovementExecutor,
 		maptos_executor: &MaptosExecutor,
 	) -> Result<(), CriterionError> {
+		// get the latest state view from the movement executor
+		let _movement_state_view = movement_executor
+			.state_view_at_version(None)
+			.map_err(|e| CriterionError::Internal(e.into()))?;
+
+		// get the latest state view from the maptos executor
+		let _maptos_state_view = maptos_executor
+			.state_view_at_version(None)
+			.map_err(|e| CriterionError::Internal(e.into()))?;
+
+		// compare the two state views
+
 		Ok(())
 	}
 }
