+++ tests/executor/util/types/Cargo.toml
@@ -16,6 +16,7 @@ maptos-opt-executor = { workspace = true }
 aptos-executor = { workspace = true }
 aptos-vm = { workspace = true }
 thiserror = { workspace = true }
+aptos-storage-interface = { workspace = true }
 
 [lints]
 workspace = true
